requirements.txt
streamlit==1.39.0
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
python-dotenv==1.0.1
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14